+++ src/pattern.h
@@ -1,6 +1,6 @@
 /*
  Sniffle
- Copyright 2018 Peter Pearson.
+ Copyright 2018-2019 Peter Pearson.
 
  Licensed under the Apache License, Version 2.0 (the "License");
  You may not use this file except in compliance with the License.
@@ -27,6 +27,7 @@ struct PatternSearch
 	enum PatternType
 	{
 		ePatternUnknown,
+		ePatternSingleFile,
 		ePatternSimple,
 		ePatternWildcardDir,
 		ePatternError